Backend/requirements.txt
fastapi
uvicorn
pandas
numpy
scikit-learn
xgboost
matminer
pymatgen
joblib
python-dotenv
pydantic